backend/app/schemas/common.py
from typing import Any, Generic, TypeVar

from pydantic import BaseModel

T = TypeVar("T")


class ApiResponse(BaseModel, Generic[T]):
    code: int = 0
    message: str = "success"
    data: T | None = None


class MessageData(BaseModel):
    message: str


class EmptyData(BaseModel):
    pass

