backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
SQLAlchemy==2.0.43
PyMySQL==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pydantic-settings==2.11.0
python-multipart==0.0.20
pytest==8.4.2
httpx==0.28.1
